Add variable with tick count in M0 port, for use with SystemView.
Define SEGGER_SYSVIEW_TickCnt for use of Segger SystemView with
Cortex-M0.

As described in Cortex-M0 section of SystemView manual, this
variable is used in place of the cycle count register present on
M3/M4 etc cores.

It must be incremented before any SYSVIEW events are generated.

It is then used by SEGGER_SYSVIEW_X_GetTimestamp to create timestamp.